+++ ir/be/bechordal.c
@@ -180,7 +180,7 @@ static ir_node *handle_constraints(be_chordal_alloc_env_t *alloc_env,
 	int *assignment;
 	pmap *partners;
 	int i, n_alloc;
-	unsigned col;
+	size_t col;
 	const ir_edge_t *edge;
 	ir_node *perm = NULL;
 	//int match_res, cost;
