+++ ir/ir/iredges.c
@@ -859,7 +859,7 @@ static void verify_edge_counter(ir_node *irn, void *env)
 	int                    list_cnt;
 	int                    ref_cnt;
 	int                    edge_cnt;
-	unsigned long          idx;
+	size_t                 idx;
 	const struct list_head *head;
 	const struct list_head *pos;
